Upgrade ndata to 0.2.15

+++ src/lib.rs
@@ -22,13 +22,13 @@ pub mod pycmd;
 use datastore::DataStore;
 use generated::Generated;
 
-pub fn init(dir:&str) -> (&str, (((usize,usize),(usize,usize)),((usize,usize),(usize,usize)),((usize,usize),(usize,usize)))) {
+pub fn init(dir:&str) -> (&str, (((u64,u64),(u64,u64)),((u64,u64),(u64,u64)),((u64,u64),(u64,u64)))) {
   let q = DataStore::init(dir);
   Generated::init();
   q
 }
 
-pub fn mirror(q:(&str, (((usize,usize),(usize,usize)),((usize,usize),(usize,usize)),((usize,usize),(usize,usize))))) {
+pub fn mirror(q:(&str, (((u64,u64),(u64,u64)),((u64,u64),(u64,u64)),((u64,u64),(u64,u64))))) {
   DataStore::mirror(q);
   Generated::init();
 }
